# console/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BlueprintLoad": {
        "additionalProperties": true,
        "description": "The 202 from asking for a swap. The work outlives the request.",
        "properties": {
          "loading": {
            "title": "Loading",
            "type": "boolean"
          },
          "run": {
            "title": "Run",
            "type": "string"
          }
        },
        "required": [
          "run",
          "loading"
        ],
        "title": "BlueprintLoad",
        "type": "object"
      },
...
      "Bucket": {
        "additionalProperties": true,
        "description": "`strategy` is null exactly when `trained` is false -- the server refuses to\nemit the uniform an allocated-but-unvisited row would otherwise read as, and\nthis keeps that distinction rather than defaulting it away.",
        "properties": {
          "reach_count": {
            "title": "Reach Count",
            "type": "integer"
          },
          "strategy": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "trained": {
            "title": "Trained",
            "type": "boolean"
          }
        },
        "required": [
          "trained",
          "strategy",
          "reach_count"
        ],
        "title": "Bucket",
        "type": "object"
      },
...
      "Edge": {
        "additionalProperties": true,
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "token",
          "type",
          "amount"
        ],
        "title": "Edge",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Hand": {
        "additionalProperties": true,
        "description": "A hand in progress.\n\n`bot_hole_cards` and every `mix` are null until `over`: the server withholds\nthem, and this says so, because a client that received them could show them\nand a sit-down where you see the opponent's hand measures nothing.",
        "properties": {
          "board": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Board",
            "type": "array"
          },
          "bot_decisions": {
            "title": "Bot Decisions",
            "type": "integer"
          },
          "bot_hole_cards": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bot Hole Cards"
          },
          "bot_untrained_decisions": {
            "title": "Bot Untrained Decisions",
            "type": "integer"
          },
          "button": {
            "title": "Button",
            "type": "integer"
          },
          "hole_cards": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Hole Cards",
            "type": "array"
          },
          "human_seat": {
            "title": "Human Seat",
            "type": "integer"
          },
          "legal": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Edge"
            },
            "title": "Legal",
            "type": "array"
          },
          "log": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HandEvent"
            },
            "title": "Log",
            "type": "array"
          },
          "over": {
            "title": "Over",
            "type": "boolean"
          },
          "payoff": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payoff"
          },
          "pot": {
            "title": "Pot",
            "type": "number"
          },
          "session": {
            "title": "Session",
            "type": "string"
          },
          "showdown": {
            "title": "Showdown",
            "type": "boolean"
          },
          "stacks": {
            "default": [],
            "items": {
              "type": "number"
            },
            "title": "Stacks",
            "type": "array"
          },
          "street": {
            "title": "Street",
            "type": "string"
          },
          "to_act": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Act"
          }
        },
        "required": [
          "session",
          "over",
          "street",
          "pot",
          "human_seat",
          "button",
          "to_act",
          "bot_hole_cards",
          "payoff",
          "showdown",
          "bot_decisions",
          "bot_untrained_decisions"
        ],
        "title": "Hand",
        "type": "object"
      },
      "HandEvent": {
        "additionalProperties": true,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "amount": {
            "title": "Amount",
            "type": "number"
          },
          "mix": {
            "anyOf": [
              {
                "items": {
                  "maxItems": 2,
                  "minItems": 2,
                  "prefixItems": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "number"
                    }
                  ],
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mix"
          },
          "seat": {
            "title": "Seat",
            "type": "integer"
          },
          "street": {
            "title": "Street",
            "type": "string"
          },
          "untrained": {
            "title": "Untrained",
            "type": "boolean"
          }
        },
        "required": [
          "seat",
          "actor",
          "action",
          "amount",
          "street",
          "untrained",
          "mix"
        ],
        "title": "HandEvent",
        "type": "object"
      },
...
      "LeftSession": {
        "additionalProperties": true,
        "description": "Confirmation that a play session was dropped on the far side.\n\nThe session lives where the blueprint does, so leaving is a request rather\nthan a local forget -- which is also why the proxy holds no state and a\nconsole restart does not lose a hand in progress.",
        "properties": {
          "dropped": {
            "title": "Dropped",
            "type": "boolean"
          },
          "session": {
            "title": "Session",
            "type": "string"
          }
        },
        "required": [
          "session",
          "dropped"
        ],
        "title": "LeftSession",
        "type": "object"
      },
...
      "NodeGrid": {
        "additionalProperties": true,
        "properties": {
          "actions": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Actions",
            "type": "array"
          },
          "actor": {
            "title": "Actor",
            "type": "integer"
          },
          "blocked": {
            "title": "Blocked",
            "type": "integer"
          },
          "board": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Board",
            "type": "array"
          },
          "buckets": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Bucket"
            },
            "default": {},
            "title": "Buckets",
            "type": "object"
          },
          "combo_buckets": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Combo Buckets",
            "type": "array"
          },
          "street": {
            "title": "Street",
            "type": "string"
          },
          "trained_buckets": {
            "title": "Trained Buckets",
            "type": "integer"
          }
        },
        "required": [
          "street",
          "actor",
          "blocked",
          "trained_buckets"
        ],
        "title": "NodeGrid",
        "type": "object"
      },
...
      "SolverNode": {
        "additionalProperties": true,
        "properties": {
          "board": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Board",
            "type": "array"
          },
          "children": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Edge"
            },
            "title": "Children",
            "type": "array"
          },
          "grid": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeGrid"
              },
              {
                "type": "null"
              }
            ]
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "terminal": {
            "title": "Terminal",
            "type": "boolean"
          }
        },
        "required": [
          "path",
          "terminal",
          "grid"
        ],
        "title": "SolverNode",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "poker-solver console",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/blueprint/load": {
      "post": {
        "operationId": "_load_api_blueprint_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintLoad"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Load"
      }
    },
    "/api/blueprint/node": {
      "get": {
        "operationId": "_node_api_blueprint_node_get",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": false,
            "schema": {
              "default": "",
              "title": "Path",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "board",
            "required": false,
            "schema": {
              "default": "",
              "title": "Board",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "average",
            "required": false,
            "schema": {
              "default": true,
              "title": "Average",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverNode"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Node"
      }
    },
    "/api/blueprint/play": {
      "post": {
        "operationId": "_start_api_blueprint_play_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hand"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Start"
      }
    },
    "/api/blueprint/play/{session_id}": {
      "delete": {
        "operationId": "_leave_api_blueprint_play__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeftSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Leave"
      },
      "get": {
        "operationId": "_hand_api_blueprint_play__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hand"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Hand"
      }
    },
    "/api/blueprint/play/{session_id}/action": {
      "post": {
        "operationId": "_act_api_blueprint_play__session_id__action_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Hand"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": " Act"
      }
    },
...
  }
}